Implement key recovery and enhance file reading in Aether KV engine
- Added RecoverKeyDir method to KVEngine for recovering key directory from file.
- Improved error handling and logging during key recovery and file operations.
- Updated ReadAt method in file.go to accept int64 for size parameter for better compatibility.
- Introduced GetKeyDirSize method to retrieve the size of the key directory.
- Enhanced decoding logic in codec.go to handle header-only records and improve error reporting.

+++ internal/engine/file.go
@@ -89,7 +89,7 @@ func (f *File) Append(data []byte) (int64, error) {
 	return offset, nil
 }
 
-func (f *File) ReadAt(offset int64, size uint32) ([]byte, error) {
+func (f *File) ReadAt(offset int64, size int64) ([]byte, error) {
 	slog.Debug("file: reading data from file",
 		"offset", offset,
 		"size", size)
